Fit second order intensity ellipses or ellipsoids

Difference to v1: Don't use the pseudo-Fry but true Fry.

Usage

fry_ellipsoids2(
  x,
  nvec = 1:5,
  r_adjust = 1,
  nangles,
  eps = 0,
  cylindric = FALSE,
  double = FALSE,
  border = TRUE,
  origin = TRUE,
  verbose = FALSE,
  keep_data = FALSE,
  data,
  ...
)

Arguments

x

point pattern

nvec

which jumps to consider

r_adjust

consider only fry points with length r_adjust*max(window side)

nangles

If 3d, number of refinements (def=2), in 2d def=20 sectors.

eps

add to the sector angle. Results in smoothing if >0 (default=0).

cylindric

Use a cylinder instead of a sector

double

Should we use both directions of a pair

border

Should border effects be cancelled by minus sampling, if TRUE, "double" ignored.

origin

Constrain the ellipses' centers to origo.

keep_data

Keep the data for each ellipsoid, inside each ellipsoids? Helps with averaging.

data

The vectors to which fit ellipses. Default is missing, and we compute Fry-points.