Inhomogeneous anisotropic pcf function, sector version

Estimate a sector/cone pcf function for second order reweighted ("inhomogeneous") pattern.

Usage

pcf_anin(
  x,
  u,
  epsilon,
  r,
  lambda = NULL,
  lambda_h,
  r_h,
  stoyan = 0.15,
  renormalise = TRUE,
  border = 1,
  divisor = "d",
  ...
)

Arguments

x

pp, list with $x~coordinates $bbox~bounding box

u

unit vector(s) of direction, as row vectors. Default: x and y axis.

epsilon

Central half angle for the directed sector/cone (total angle of the rotation cone is 2*epsilon)

r

radius vector at which to evaluate the function

lambda

optional vector of intensity estimates at points

lambda_h

if lambda missing, use this bandwidth in a kernel estimate of lambda(x)

r_h

smoothing for range dimension, epanechnikov kernel

stoyan

If r_h not given, use r_h=stoyan/lambda^(1/dim). Same as 'stoyan' in spatstat's pcf.

renormalise

See details.

border

Use translation correction? Default=1, yes. Only for cuboidal windows.

divisor

See spatstat's pcf.ppp for this.

...

passed on to e.g. intensity_at_points

Value

Returns a dataframe.

Details

Computes a second order reweighted version of the sector-pcf. The sector pcf differs from the true anisotropic pcf by assuming that the anisotropic pcf is constant over the small arc/cap of the sector, thus averaging over that data-area and providing more stable estimates.

Lambda(x) at points can be given, or else it will be estimated using kernel smoothing. See intensity_at_points.

If 'renormalise=TRUE', we normalise the lambda estimate so that sum(1/lambda(x))=|W|. This corresponds in spatstat's Kinhom to setting 'normpower=2'.

See also